+++ bcbio/ngsalign/novoalign.py
@@ -2,15 +2,13 @@
 
 For BAM input handling this requires:
   novoalign (with license for multicore)
-  novosort (also with license for multicore)
   samtools
 """
 import os
 import subprocess
 
-from bcbio import utils
-from bcbio.distributed.transaction import file_transaction
-from bcbio.ngsalign import alignprep
+from bcbio import bam, utils
+from bcbio.ngsalign import alignprep, postalign
 from bcbio.pipeline import config_utils
 from bcbio.provenance import do
 from bcbio.utils import (memoize_outfile, file_exists, curdir_tmpdir)
@@ -20,17 +18,11 @@ from bcbio.utils import (memoize_outfile, file_exists, curdir_tmpdir)
 def get_rg_info(names):
     return r"@RG\tID:{rg}\tPL:{pl}\tPU:{pu}\tSM:{sample}".format(**names)
 
-def align_bam(in_bam, ref_file, names, align_dir, config):
-    """Perform realignment of input BAM file, handling sorting of input/output with novosort.
-
-    Uses unix pipes for avoid IO writing between steps:
-      - novosort of input BAM to coordinates
-      - alignment with novoalign
-      - conversion to BAM with samtools
-      - coordinate sorting with novosort
+def align_bam(in_bam, ref_file, names, align_dir, data):
+    """Perform realignment of input BAM file; uses unix pipes for avoid IO.
     """
+    config = data["config"]
     out_file = os.path.join(align_dir, "{0}-sort.bam".format(names["lane"]))
-    novosort = config_utils.get_program("novosort", config)
     novoalign = config_utils.get_program("novoalign", config)
     samtools = config_utils.get_program("samtools", config)
     resources = config_utils.get_resources("novoalign", config)
@@ -40,16 +32,14 @@ def align_bam(in_bam, ref_file, names, align_dir, config):
 
     if not file_exists(out_file):
         with curdir_tmpdir(base_dir=align_dir) as work_dir:
-            with file_transaction(out_file) as tx_out_file:
+            with postalign.tobam_cl(data, out_file, bam.is_paired(in_bam)) as (tobam_cl, tx_out_file):
                 rg_info = get_rg_info(names)
-                cmd = ("{novosort} -c {num_cores} -m {max_mem} --compression 0 "
-                       " -n -t {work_dir} {in_bam} "
+                tx_out_prefix = os.path.splitext(tx_out_file)[0]
+                prefix1 = "%s-in1" % tx_out_prefix
+                cmd = ("{samtools} sort -n -o -l 0 -@ {num_cores} -m {max_mem} {in_bam} {prefix1} "
                        "| {novoalign} -o SAM '{rg_info}' -d {ref_file} -f /dev/stdin "
-                       "  -F BAMPE -c {num_cores} {extra_novo_args} "
-                       "| {samtools} view -b -S -u - "
-                       "| {novosort} -c {num_cores} -m {max_mem} -t {work_dir} "
-                       "  -o {tx_out_file} /dev/stdin")
-                cmd = cmd.format(**locals())
+                       "  -F BAMPE -c {num_cores} {extra_novo_args} | ")
+                cmd = cmd.format(**locals()) + tobam_cl
                 do.run(cmd, "Novoalign: %s" % names["sample"], None,
                        [do.file_nonempty(tx_out_file), do.file_reasonable_size(tx_out_file, in_bam)])
     return out_file
@@ -78,13 +68,11 @@ def align_pipe(fastq_file, pair_file, ref_file, names, align_dir, data):
     rg_info = get_rg_info(names)
     if not utils.file_exists(out_file) and (final_file is None or not utils.file_exists(final_file)):
         with utils.curdir_tmpdir() as work_dir:
-            with file_transaction(out_file) as tx_out_file:
+            with postalign.tobam_cl(data, out_file, pair_file != "") as (tobam_cl, tx_out_file):
                 tx_out_prefix = os.path.splitext(tx_out_file)[0]
                 cmd = ("{novoalign} -o SAM '{rg_info}' -d {ref_file} -f {fastq_file} {pair_file} "
-                       "  -c {num_cores} {extra_novo_args} "
-                       "| {samtools} view -b -S -u - "
-                       "| {samtools} sort -@ {num_cores} -m {max_mem} - {tx_out_prefix}")
-                cmd = cmd.format(**locals())
+                       "  -c {num_cores} {extra_novo_args} | ")
+                cmd = cmd.format(**locals()) + tobam_cl
                 do.run(cmd, "Novoalign: %s" % names["sample"], None,
                        [do.file_nonempty(tx_out_file), do.file_reasonable_size(tx_out_file, fastq_file)])
     data["work_bam"] = out_file
@@ -118,29 +106,6 @@ def _novoalign_args_from_config(config, need_quality=True):
 # Tweaks to add
 # -k -t 200 -K quality calibration metrics
 # paired end sizes
-
-def align(fastq_file, pair_file, ref_file, out_base, align_dir, data,
-          extra_args=None, names=None):
-    """Align with novoalign.
-    """
-    raise NotImplementedError("Prefer align_pipe approach now.")
-    config = data["config"]
-    rg_name = names.get("rg", None) if names else None
-    out_file = os.path.join(align_dir, "{0}.sam".format(out_base))
-    if not file_exists(out_file):
-        cl = [config_utils.get_program("novoalign", config)]
-        cl += _novoalign_args_from_config(config)
-        cl += extra_args if extra_args is not None else []
-        cl += ["-o", "SAM"]
-        if rg_name:
-            cl.append(r"'@RG\tID:{0}'".format(rg_name))
-        cl += ["-d", ref_file, "-f", fastq_file]
-        if pair_file:
-            cl.append(pair_file)
-        with file_transaction(out_file) as tx_out_file:
-            cmd = "{cl} > {out_file}".format(cl=" ".join(cl), out_file=tx_out_file)
-            do.run(cmd, "novoalign {rg_name}".format(**locals()), None)
-    return out_file
 
 # ## Indexing