+++ bcbio/pipeline/alignment.py
@@ -5,12 +5,9 @@ This works as part of the lane/flowcell process step of the pipeline.
 from collections import namedtuple
 import os
 
-from Bio.SeqIO.QualityIO import FastqGeneralIterator
-
 from bcbio import bam, utils
 from bcbio.bam import cram
-from bcbio.distributed.transaction import file_transaction
-from bcbio.ngsalign import (bowtie, bwa, tophat, bowtie2, mosaik,
+from bcbio.ngsalign import (bowtie, bwa, tophat, bowtie2,
                             novoalign, star)
 
 # Define a next-generation sequencing tool to plugin:
@@ -33,7 +30,7 @@ TOOLS = {
     "bowtie2": NgsTool(bowtie2.align, None,
                        bowtie2.galaxy_location_file, bowtie2.remap_index_fn),
     "bwa": NgsTool(bwa.align_pipe, bwa.align_bam, bwa.galaxy_location_file, None),
-    "mosaik": NgsTool(mosaik.align, None, mosaik.galaxy_location_file, None),
+#    "mosaik": NgsTool(mosaik.align, None, mosaik.galaxy_location_file, None),
     "novoalign": NgsTool(novoalign.align_pipe, novoalign.align_bam,
                          novoalign.galaxy_location_file, novoalign.remap_index_fn),
     "tophat": NgsTool(tophat.align, None,
@@ -85,7 +82,7 @@ def _align_from_bam(fastq1, aligner, align_ref, sam_ref, names, align_dir, data)
     align_fn = TOOLS[aligner].bam_align_fn
     if align_fn is None:
         raise NotImplementedError("Do not yet support BAM alignment with %s" % aligner)
-    return align_fn(fastq1, align_ref, names, align_dir, config)
+    return align_fn(fastq1, align_ref, names, align_dir, data)
 
 def _align_from_fastq(fastq1, fastq2, aligner, align_ref, sam_ref, names,
                       align_dir, data):
@@ -103,32 +100,3 @@ def _align_from_fastq(fastq1, fastq2, aligner, align_ref, sam_ref, names,
         work_bam = bam.sam_to_bam(out, config)
         data["work_bam"] = bam.sort(work_bam, config)
         return data
-
-def _remove_read_number(in_file, sam_file):
-    """Work around problem with MergeBamAlignment with BWA and single end reads.
-
-    Need to remove read number ends from Fastq to match BWA stripping of numbers.
-
-    http://sourceforge.net/mailarchive/forum.php?thread_name=87bosvbbqz.fsf%
-    40fastmail.fm&forum_name=samtools-help
-    http://sourceforge.net/mailarchive/forum.php?thread_name=4EB03C42.2060405%
-    40broadinstitute.org&forum_name=samtools-help
-    """
-    out_file = os.path.join(os.path.dirname(sam_file),
-                            "%s-safe%s" % os.path.splitext(os.path.basename(in_file)))
-    # file already exists and is zero means we already skipped the removal and
-    # are just using the original file
-    if os.path.exists(out_file) and os.path.getsize(out_file) == 0:
-        return in_file
-    if not os.path.exists(out_file):
-        with file_transaction(out_file) as tx_out_file:
-            with open(in_file) as in_handle:
-                with open(tx_out_file, "w") as out_handle:
-                    for i, (name, seq, qual) in enumerate(FastqGeneralIterator(in_handle)):
-                        if i == 0 and not name.endswith("/1"):
-                            out_file = in_file
-                            break
-                        else:
-                            name = name.rsplit("/", 1)[0]
-                            out_handle.write("@%s\n%s\n+\n%s\n" % (name, seq, qual))
-    return out_file